pyproject.toml
[tool.poetry]
name = "freeze-poetry"
version = "0.1.0"
description = ""
authors = ["Michael Van Wickle <mvanwickle@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
attrs = "^19.3.0"
click = "^7.0"
django = {version = ">=2.0", optional = true}
funcy = "^1.13"
toml = "^0.10.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.7.8"
isort = "^4.3.21"
pytest = "^3.0"

[tool.poetry.extras]
django = ["django"]


[tool.isort]
known_first_party=[
    "freeze_poetry",
]
no_lines_before=["LOCALFOLDER"]
multi_line_output=3
include_trailing_comma=true
force_grid_wrap=0
use_parentheses=true
line_length=88

[tool.poetry.scripts]
perf_parse = "freeze_poetry.__main__:cli"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
